Created new agent (v1_toying) and already submitted it to the competition (score: 600).
Created new agent in v1_toying folder. This agent is very simple; workers gather the nearest resource until full, then either take it to an existing CityTile, if that city does not have enough fuel to last the night, or builds a new city. CityTiles create the max number of workers, then research otherwise. No carts are build and the agent is totally oblivious to its competition.

Some of the collision code was inspired by sentdex's solution.